--- src/main/java/boombimapi/global/geo/core/Clusterer.java
@@ -1,16 +1,0 @@
-package boombimapi.global.geo.core;
-
-import java.util.List;
-
-public interface Clusterer {
-
-    List<ClusterMarker> cluster(
-        List<ClusterPoint> clusterPoints,
-        double minLatitude,
-        double maxLatitude,
-        double minLongitude,
-        double maxLongitude,
-        int zoomLevel
-    );
-
-}
